% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_geodb.R
\name{geodata_overview}
\alias{geodata_overview}
\title{overview of geodatabase}
\usage{
geodata_overview(all = FALSE)
}
\arguments{
\item{all}{see the whole meta data table}
}
\description{
Get the overview of the geodata in the database. You can decide which column you want to use for subsetting the data.
}
\examples{
\dontrun{
geodata_overview()
}
}
