% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_geodb.R
\name{source_geodb}
\alias{source_geodb}
\title{read database credentials}
\usage{
source_geodb(path)
}
\arguments{
\item{path}{the path to source file}
}
\description{
read database credentials
}
\examples{
\dontrun{
source_geodb()
}
}
