% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_geodb.R
\name{con_geodb}
\alias{con_geodb}
\title{connect to database
Provides a robust and flexible database connection utility
that automatically manages the geodata-base connection between internal and external
database hosts.}
\usage{
con_geodb(intern = TRUE)
}
\description{
connect to database
Provides a robust and flexible database connection utility
that automatically manages the geodata-base connection between internal and external
database hosts.
}
\examples{
\dontrun{
con_geodb()
}
}
