% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_geodb.R
\name{con_geodb}
\alias{con_geodb}
\title{connect to database
So not change the credentials. This will automatically filled by the db_source file.}
\usage{
con_geodb(
  db,
  host_db,
  db_port,
  db_user,
  db_password,
  host_db_ex,
  db_port_ex,
  intern = TRUE
)
}
\arguments{
\item{db}{name of database}

\item{host_db}{host name}

\item{db_port}{port}

\item{db_user}{user_name}

\item{db_password}{password}

\item{host_db_ex}{external hostname}

\item{db_port_ex}{external port}

\item{intern}{are you in the izw network?}
}
\description{
connect to database
So not change the credentials. This will automatically filled by the db_source file.
}
\examples{
\dontrun{
con_geodb()
}
}
