% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_geodb.R
\name{con_geodb}
\alias{con_geodb}
\title{connect to database
So not change the credentials. This will automatically filled by the db_source file.}
\usage{
con_geodb(intern = TRUE)
}
\arguments{
\item{intern}{are you in the izw network?}
}
\description{
connect to database
So not change the credentials. This will automatically filled by the db_source file.
}
\examples{
\dontrun{
con_geodb()
}
}
